refactor(dispatch): extract envelope_adapters_claude.py leaf module (PR-5 monolith split)
Moves CodexAdapter and ClaudeSubprocessAdapter out of dispatch_envelope.py
unchanged. _AdapterResult comes from envelope_types.py (PR-1), not the
facade. dispatch_envelope.py re-exports both classes at bare name so all
nine attribute-form couplings (patch.object/monkeypatch.setattr on the
class object, plus bare-name imports) keep binding — verified via an
explicit identity check (dispatch_envelope.CodexAdapter is
envelope_adapters_claude.CodexAdapter, same for ClaudeSubprocessAdapter).

Census expectation file regenerated: family gains envelope_adapters_claude,
layer4_classes/layer4_functions move their owner label for the two classes;
the couplings list itself is unchanged (test files import from
dispatch_envelope, which still resolves through the facade re-export).

Dispatch-ID: 20260804-003531-pr5-envelope-adapters-claude
